Change namespace and reorganise file locations.
