+++ app/src/main/java/com/difo/motoactv_export/DropboxUpdate.java
@@ -1,5 +1,5 @@
 
-package com.sdsoft.motoactv_export;
+package com.difo.motoactv_export;
 
 /*
  * SDSOFT Motoactv Exporter
