+++ app/src/main/java/com/difo/motoactv_export/RunKeeperObjects.java
@@ -1,4 +1,4 @@
-package com.sdsoft.motoactv_export;
+package com.difo.motoactv_export;
 
 /*
  * SDSOFT Motoactv Exporter
